lib/sampler: Add new library for sampling stats
RFC v2:
- Use linked lists for sink/source
- Remove unnecessary error handling
- Add interval/duration handling
- Fixed wrong rti_* prefixes for file names
- Header file split into session, source and sink
- Compiles now :)

Storage for ids, values for sources are still in the core library.
And they are stored per source, not per session. This makes it
simpler to provide an array of ids to the source for collection.
But it makes complicated to report the stats to sink, and needs
to be done per source and neccesitates report_begin/report_append/
report_end APIs for the sink. Needs to be reworked, will come back
with a new version soon.